cabal: explicitly fail build on unsupported platforms
Prior to this change, since an unsupported platform would trigger a
'build-depends' of 'unknown-vty-build-platform', that meant a malicious
community user could upload a package by that name containing a
malicious re-implementation of an existing Vty backend. The original
motivation for doing that was to make build failures somewhat
understandable. This change just marks the build as unbuildable in those
cases rather than run the security risk.